requirements.txt
wheel 
Pillow
djangorestframework
djangorestframework-simplejwt

drf-yasg
psycopg2-binary
python-decouple
python-slugify